pyproject.toml
[project]
name = "tertius"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "bookman",
    "pyzmq>=26.0.0",
]

[tool.uv.sources]
orbis = { path = "../orbis", editable = true }
bookman = { path = "../sand", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hypothesis>=6.152.1",
    "orbis",
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "vulture",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tertius"]

[tool.ruff]
line-length = 130

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "BLE", "SIM", "RUF"]

[tool.mypy]
strict = true
